requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart
python-dotenv
langchain==0.1.9
langchain-core==0.1.41
langchain-community==0.0.13
langchain-openai==0.0.6
langgraph==0.2.53   